% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{CalcDissimilarity}
\alias{CalcDissimilarity}
\title{Calculation of repertoire dissimilarities}
\usage{
CalcDissimilarity(
  x,
  level = c("aaClone", "ntClone", "V", "J", "VJ", "ntCDR3", "aaCDR3"),
  method = c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski",
    "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial",
    "chao", "cao", "mahalanobis"),
  binary = FALSE
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire on which the indices are computed. Should be one of "aaClone","ntClone", "V", "J", "VJ", "ntCDR3" or "aaCDR3".}

\item{method}{a character specifying the distance method to be computed. Should be one of the following: "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis."}

\item{binary}{a boolean indicating whether or not to transform the data into a presence/absence data. Default is FALSE}
}
\description{
This function assesses pairwise repertoire dissimilarities using a specific dissimilarity method.

It calculates a list of dissimilarity indices, each taking into account different parameters. The proposed methods include:

 The Jaccard similarity: a measure of similarity between sample sets defined as the size of the intersection divided by the size of the union of the sample sets.

 The Morisita-Horn similarity: a measure of similarity that tends to be over-sensitive to abundant species.
}
\details{
Details on the calculated indices can be found in the vegan package: 
https://www.rdocumentation.org/packages/vegan/versions/2.4-2/topics/vegdist
}
\examples{


CalcDissimilarity(x = RepSeqData, 
                  level = "V", method = "jaccard" )

}
