% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{chaoWor}
\alias{chaoWor}
\title{Adjusted Chao1}
\usage{
chaoWor(x)
}
\arguments{
\item{x}{a vector of count. Can be the list clonotype counts within a given sample.}
}
\value{
the computed Chao1 value
}
\description{
calculate the adjusted Chao1 index
}
\details{
function computes the adjusted Chao1 index for sampling without replacement
}
\examples{
set.seed(1234)
x <- rbinom(20, 5, 0.5)
chaoWor(x)
}
\keyword{internal}
