% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{diffExpInd}
\alias{diffExpInd}
\title{Differential expression analysis between two samples}
\usage{
diffExpInd(
  x,
  sampleNames = NULL,
  level = c("clone", "clonotype", "V", "J", "VJ", "CDR3nt", "CDR3aa"),
  scale = c("frequency", "count"),
  th = 1.5,
  remove.zeros = TRUE
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{sampleNames}{a vector of character with the sample_ids of the repertoires to drop from the RepSeqExperiment object.}

\item{level}{a character specifying the repertoire level to be analyzed. Should be one of "clone","clonotype","CDR3aa","CDR3nt","V", "J",or "VJ".}

\item{scale}{a character specifying the type of occurrence to take into account: "count" or "frequency".}

\item{th}{the lof2FC threshold to be used}

\item{remove.zeros}{a boolean indicating whether or not repertoire levels that are completely absent in one of the two compared samples should be to take into account in the calculation.}
}
\value{
a data frame containing the list of differentially expressed repertoire levels, their occurrence in both samples and their calculated log2FC.
}
\description{
This function compares the expression of a repertoire level between two samples.
Log-ratios are calculated on the occurrence of a selected repertoire level between two compared samples, and differentially expressed genes/sequences are identified based on the user-defined log-ratio threshold.
}
\examples{

data(RepSeqData)

diff_expression <- diffExpInd(RepSeqData, 
                              level="V", 
                              scale="frequency", 
                              sampleNames = c("tripod-30-813_R1","tripod-30-815_R1"), 
                              remove.zeros = FALSE)

}
