% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{divLevel}
\alias{divLevel}
\title{Diversity index}
\usage{
divLevel(
  x,
  index = c("shannon", "simpson", "invsimpson"),
  level = c("VJ", "V", "J"),
  norm = TRUE,
  base = exp(1)
)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{RepSeqExperiment}}]}

\item{index}{character, the diversity index to calculate. Should be one of "chao1.se", "chaowor","shannon","invsimpson","simpson" or "gini".}

\item{level}{character, the level of the repertoire to estimate. Should be one of "V", "J" or "VJ".}

\item{norm}{boolean, whether to compute the normalized diversity index, which divides the computed value by the log(S), S being the number of observed species in the repetoire. Default is TRUE.}

\item{base}{a positive or complex number, the base with respect to which logarithms are computed. Default is e=exp(1).}
}
\value{
a data.table
}
\description{
calculate a sublist of diversity indices on gene segments.
}
\details{
function computes the diversity indices for level \code{VJ}, \code{V} or \code{J} only.
}
\examples{

data(RepSeqData)

divLevel(x = RepSeqData, index = "shannon", level = "V", norm = TRUE)

divLevel(x = RepSeqData, index = "simpson", level = "J", norm = FALSE)

divLevel(x = RepSeqData, index = "invsimpson", level = "VJ", norm = FALSE)

}
\keyword{internal}
