% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{.safe_wilcox}
\alias{.safe_wilcox}
\title{Calculate wilcoxon safe}
\usage{
.safe_wilcox(df, group_var, facet_var, color_var, value_var)
}
\arguments{
\item{df}{dataframe}

\item{group_var}{x axis}

\item{facet_var}{facet group(s)}

\item{color_var}{color group}
}
\value{
dataframe of statistical test results
}
\description{
calculate wilcoxon tests in all plots
}
\details{
function calculates wilcoxon tests in all plots
}
\keyword{internal}
