% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{.toDESeq2}
\alias{.toDESeq2}
\title{Differential expression analysis}
\usage{
.toDESeq2(
  x,
  colGrp,
  level = c("aaClone", "ntClone", "V", "J", "VJ", "ntCDR3", "aaCDR3")
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{colGrp}{a vector of character, specify the groups to be compared.}

\item{level}{character, the level of the repertoire to estimate. Should be one of "aaClone", "V", "J", "VJ" or "aaCDR3".}
}
\value{
a DESeqDataSet object
}
\description{
Differential expression analysis using DESeq2
}
\details{
function compares the expression level of a gene segment or a clone using the DESEq2 package.
}
\keyword{internal}
