% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filetype}
\alias{filetype}
\title{Give type of file}
\usage{
filetype(path)
}
\arguments{
\item{path}{path to a file}
}
\value{
a string indicating file type
}
\description{
get the type of file
}
\details{
This function allows to get the type of a file
}
\examples{

l <- list.files(system.file(file.path('extdata/MiAIRR'),
                     package = 'AnalyzAIRR'),
                     full.names = TRUE)
path <- l[1]

file_type <- filetype(path)

}
\keyword{internal}
