% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filterCount}
\alias{filterCount}
\title{Filtering based on sequence count}
\usage{
filterCount(
  x,
  level = c("aaClone", "ntClone", "aaCDR3", "ntCDR3"),
  n = 1,
  group = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the type of sequences on which the
filtering will be applied. Should be one of "aaClone","ntClone","aaCDR3" or
"ntCDR3". For instance, for level="aaCDR3", counts will first be recalculated
based on this column. Then, aaCDR3 sequences with counts equal or below the
n threshold will be excluded.}

\item{n}{an integer specifying the count threshold below which sequences will
be filtered out. For instance, for n=2, sequences with a count of 1 and 2
are filtered out.}

\item{group}{a vector of character indicating the group column name in mData
and one experimental group within this column. Samples belonging to the
chosen group will be analyzed. The column must be of class factor. Default
is NULL, values are calculated in all the samples within the dataset.}
}
\value{
an object of class \code{RepSeqExperiment}
}
\description{
This function filters out, in each sample, sequences having a
count equal to or below a specified threshold. It can be applied on all the
samples within a RepSeqExperiment object, or a group of samples belonging to
a specific experimental group.
}
\examples{

filterdata <- filterCount(x = RepSeqData,
                          n = 1,
                          level = "aaCDR3",
                          group = c("sex", "M"))


}
