% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filterSequence}
\alias{filterSequence}
\title{Exclude a sequence from a RepSeqExperiment object.}
\usage{
filterSequence(
  x,
  level = c("aaClone", "ntClone", "aaCDR3", "ntCDR3"),
  name,
  group = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken
into account. Should be one of "aaClone","ntClone", "ntCDR3" or "aaCDR3".}

\item{name}{a vector of character specifying the name(s) of the sequence(s)
to filter out from the RepSeqExperiment object.}

\item{group}{a vector of character indicating the group column name in the
mData slot and one experimental group within this column.}
}
\value{
a RepSeqExperiment object.
}
\description{
This function allows to drop of one or several sequences from a
RepSeqExperiment object in all samples or a specified group of samples.
}
\examples{

data(RepSeqData)

RepSeqData<- filterSequence(x = RepSeqData,
                            level="aaClone",
                            name="TRAV11 CVVGDRGSALGRLHF TRAJ18",
                            group=c("cell_subset","Teff"))

}
