% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{formatSingleCell}
\alias{formatSingleCell}
\title{A method for the building of a RepSeqExperiment object using single cell data}
\usage{
formatSingleCell(path)
}
\arguments{
\item{path}{a path to the acell ranger filtered contigs file}
}
\value{
an object of class \code{\link{RepSeqExperiment}}. See \code{\link{RepSeqExperiment-class}} for more details.
}
\description{
This function
}
\keyword{internal}
