% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getPrivate}
\alias{getPrivate}
\title{Extract private sequences}
\usage{
getPrivate(
  x,
  level = c("ntCDR3", "aaCDR3", "ntClone", "aaClone"),
  singletons = FALSE
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken
into account. Should be one of aaClone","ntClone", "ntCDR3" or "aaCDR3".}

\item{singletons}{a boolean indicating whether or not private sequences with
a count of 1 should be extracted. Default is FALSE.}
}
\value{
an object of class \code{\linkS4class{RepSeqExperiment}} composed
exclusively of private sequences.
}
\description{
This function extract private sequences from the whole dataset,
 i.e. sequences found exclusively in one sample.
}
\examples{

data(RepSeqData)

privateclones <- getPrivate(RepSeqData,
                            level = "ntClone",
                            singletons = FALSE)

}
