% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getProductive}
\alias{getProductive}
\title{Extract productive sequences}
\usage{
getProductive(x)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}.}
}
\value{
a filtered \code{\linkS4class{RepSeqExperiment}} object exclusively containing productive sequences.
}
\description{
Extract productive sequences from a RepSeqExperiment object by filtering out unproductive ones. Filtered sequences include:

- out-of-frame sequences: sequences with frame shifts based on the number of nucleotides in the CDR3nt column.

- sequences containing stop codons: CDR3aa sequences with a "*" or "~" symbols.
}
\examples{

data(RepSeqData)
productiveData <- getProductive(x = RepSeqData)

}
