% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getTopSequences}
\alias{getTopSequences}
\title{Extract the most frequent sequences}
\usage{
getTopSequences(
  x,
  level = c("aaClone", "ntClone", "aaCDR3", "ntCDR3"),
  group = NULL,
  prop = 0.01
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken
into account. Should be one of aaClone","ntClone", "ntCDR3" or "aaCDR3".}

\item{group}{character, column name in mData indicating the group on which
the extraction will be applied. Must be of class factor. Default is NULL.}

\item{prop}{a numeric between 0 and 1 indicating the proportion of top
sequences to extract.}
}
\value{
an object of class \code{\linkS4class{RepSeqExperiment}}
}
\description{
This function extracts the top n sequences within all samples
or a group of samples.
}
\examples{

data(RepSeqData)
topClones <- getTopSequences(x = RepSeqData,
                          level = "aaClone",
                          group = c("cell_subset", "amTreg"), prop = 0.1)


}
