% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getUnproductive}
\alias{getUnproductive}
\title{Extract unproductive sequences}
\usage{
getUnproductive(x)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}.}
}
\value{
a filtered \code{\linkS4class{RepSeqExperiment}} object exclusively
containing unproductive sequences.
}
\description{
Extract unproductive sequences from a RepSeqExperiment object,
which include:

- out-of-frame sequences: sequences with frame shifts based on the number of
nucleotides in the ntCDR3 column

- sequences containing stop codons: aaCDR3s with a "*" or "~" symbols.
}
\examples{

\donttest{
data(RepSeqData)

unproductiveData <- getUnproductive(x = RepSeqData)

}

}
