% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityFunctions.R
\name{iChao}
\alias{iChao}
\title{Improved Chao1}
\usage{
iChao(x)
}
\arguments{
\item{x}{a vector of count. Can be the list clonotype counts within a given sample.}
}
\value{
the computed iChao1 value
}
\description{
calculate the improved Chao1 index
}
\details{
function computes the improved version of Chao1, Chao and Lin Chao, A. and Lin, C.-W. (2012).
}
\examples{
set.seed(1234)
x <- rbinom(20, 5, 0.5)
iChao(x)
}
\keyword{internal}
