% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{Intervals}
\alias{Intervals}
\title{Calculation of the clonal distribution per interval}
\usage{
Intervals(
  x,
  level = c("aaClone", "ntClone", "ntCDR3", "aaCDR3"),
  fractions = c("count", "frequency")
)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken into account when calculating the clonal distribution. Should be one of "aaClone","ntClone", "ntCDR3" or "aaCDR3".}

\item{fractions}{whether intervals should be determined in count or frequency}
}
\description{
This function calculates the clonal distribution per a set of intervals in all the samples within the dataset. 

The column titled "Distribution" calculates the proportion of each interval in the whole repertoire, whereas the one titled "Cumulative frequency" shows the cumulative frequency of the sequences within each interval.

This could allow a global view of the repertoire fraction contributing the most to the repertoire. For instance, top sequences belonging to the highest interval often constitute a low fraction of the whole repertoire but contribute more significantly in terms of cumulative frequency in view of their high occurrence.
}
\examples{

data(RepSeqData)

Intervals(x = RepSeqData, level="aaCDR3",  fractions="count")

}
