% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{mergeRepSeq}
\alias{mergeRepSeq}
\title{Merge RepSeqExperiment objects}
\usage{
mergeRepSeq(a, b)
}
\arguments{
\item{a}{the first  \code{\linkS4class{RepSeqExperiment}} object.}

\item{b}{the second  \code{\linkS4class{RepSeqExperiment}} object.}
}
\value{
a  \code{\linkS4class{RepSeqExperiment}} object containing all
information from the two merged objects.
}
\description{
This function allows to merge of two RepSeqExperiement objects.

The two objects must however contain unique and different sample_ids.

This function can be useful in case users wish to analyze alignment files
from different formats, or biological/technical replicates.
}
