% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{named.contr.sum}
\alias{named.contr.sum}
\title{Contrast matrix creation}
\usage{
named.contr.sum(x, ...)
}
\arguments{
\item{x}{a factor}

\item{...}{other parameters will be passed to function \code{contr.sum}}
}
\value{
a contrast design matrix
}
\description{
create contrast matrix from factor level names
}
\details{
function creates contrast matrix from factor level names.
}
\keyword{internal}
