% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{perturbationScore}
\alias{perturbationScore}
\title{Calculation of the perturbation score}
\usage{
perturbationScore(
  x,
  ctrl.names,
  distance = c("manhattan", "euclidean", "canberra", "minkowski", "maximum"),
  p = 2
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{ctrl.names}{a vector of characters indicating the sample_ids to be used as controls.}

\item{distance}{a character specifying the distance method to be used in the calculation of the perturbation scores. Should be one of the following: "manhattan", "euclidean", "canberra", "minkowski" or "maximum".}

\item{p}{an integer indicating the power of the Minkowski distance. Default is 2.}
}
\value{
a data frame containing the perturbation scores for each V-gene.
}
\description{
This function computes the perturbation scores of the CDR3aa length distribution within each V gene.

Scores are calculated as a distance between each repertoire and the mean of the control group using the ISEApeaks method (Colette and Six., 2002).
}
\examples{

data(RepSeqData)
pert <- perturbationScore(RepSeqData,
                          ctrl.names = c("tripod-30-813_R1",
                                         "tripod-30-815_R1",
                                         "tripod-31-846_R1"),
                          distance = "manhattan")

}
