% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{plotColors}
\alias{plotColors}
\title{Color palette}
\usage{
plotColors(x)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}
}
\value{
a list of distinct colors assigned to each sample/group in the mData.
}
\description{
This function allows an automatic color assigning, chosen from
color-blinded friendly palettes, to every sample and experimental group in
the metadata.
As such, each sample/group will be assigned the same color in all the
visualization functions.
}
\examples{

data(RepSeqData)

colors <- plotColors(x = RepSeqData)

}
