% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDiffExp}
\alias{plotDiffExp}
\title{Visualization of differential expression in a volcano plot}
\usage{
plotDiffExp(
  x,
  level = c("clone", "clonotype", "V", "J", "VJ", "CDR3nt", "CDR3aa"),
  group = c("cell_subset", "amTreg", "nTreg"),
  plot = c("Volcano", "PCA"),
  label_colors = NULL,
  FC.TH = 2,
  PV.TH = 0.05,
  top = 10
)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire on which the diversity should be estimated. Should be one of "clone","clonotype", "V", "J", "VJ", "CDR3nt" or "CDR3aa".}

\item{group}{a vector of character indicating the column name in the mData slot, as well as the two groups to be compared.}

\item{plot}{a character indicating the type of visualization in which the results will be represented, either a volcano plot or a PCA.}

\item{label_colors}{a list of colors for each factor column in metaData. See \code{\link{plotColors}}. If NULL, default colors are used. Only used when plot is set to "PCA".}

\item{FC.TH}{an integer indicating the log2FoldChange threshold. Default is 2. Only required when plot is set to "Volcano".}

\item{PV.TH}{an integer indicating the adjusted pvalue threshold. Default is 0.05. Only required when plot is set to "Volcano".}

\item{top}{an integer indicating the top n significant labels to be shown on the volcano plot. Default is 10. Only required when plot is set to "Volcano".}
}
\description{
This function plots differentially expressed repertoire levels calculated using the \code{\link{diffExpGroup}} in a volcano plot.
}
\examples{

data(RepSeqData)
plotDiffExp(x = RepSeqData,
            level = "V",
            group = c("cell_subset", "amTreg", "nTreg"),
            plot="Volcano",
            top = 10,
            FC.TH = 1,
            PV.TH = 0.05)

plotDiffExp(x = RepSeqData,
            level = "V",
            group = c("cell_subset", "amTreg", "nTreg"),
            plot="PCA",
            label_colors=NULL)
            
}
