% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDimReduction}
\alias{plotDimReduction}
\title{Multidimensional Scaling analysis and Principal Component Analysis}
\usage{
plotDimReduction(
  x,
  level = c("clone", "clonotype", "V", "J", "VJ", "CDR3nt", "CDR3aa"),
  method = c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski",
    "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial",
    "chao", "cao", "mahalanobis"),
  colorBy = NULL,
  label_colors = NULL,
  dim_method = c("MDS", "PCA")
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire on which the diversity should be estimated. Should be one of "clone","clonotype", "V", "J", "VJ", "CDR3nt" or "CDR3aa".}

\item{method}{a character specifying the distance method to be computed. Should be specified only if the dim_method parameter is set to "MDS", and should be one of the following: "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis."}

\item{colorBy}{a character specifying the column name in mData to be used to attribute group colors.}

\item{label_colors}{a list of colors for each factor column in metaData. See \code{\link{plotColors}}. If NULL, default colors are used.}

\item{dim_method}{a character indicating the dimensional reduction method to be performed. Should be one of "PCA" or "MDS".}
}
\description{
This function can be used to visualize:

- repertoire dissimilarities by performing a multidimensional scaling (MDS) on a pairwise distance matrix calculated between samples on a selected repertoire level, using a specific dissimilarity method.
 The proposed methods are detailed in the \code{\link{plotDissimilarityMatrix}} function.

- differentially expressed repertoire levels calculated using the \code{\link{diffExpGroup}} function by performing a principal component analysis (PCA).
}
\details{
Details on the proposed dissimilarity indices can be found in the vegan package:

https://www.rdocumentation.org/packages/vegan/versions/2.4-2/topics/vegdist
}
\examples{

data(RepSeqData)

plotDimReduction(x = RepSeqData,
                 level = "V",
                 method = "euclidean",
                 colorBy = "cell_subset",
                 dim_method="MDS")

plotDimReduction(x = RepSeqData,
                level = "J",
                colorBy = "cell_subset",
                dim_method="PCA")

}
