% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDissimilarityMatrix}
\alias{plotDissimilarityMatrix}
\title{Visualization of repertoire dissimilarities}
\usage{
plotDissimilarityMatrix(
  x,
  level = c("clone", "clonotype", "V", "J", "VJ", "CDR3nt", "CDR3aa"),
  method = c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski",
    "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial",
    "chao", "cao", "mahalanobis"),
  clustering = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  binary = FALSE,
  label_colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire on which the indices are computed. Should be one of "clone","clonotype", "V", "J", "VJ", "CDR3nt" or "CDR3aa".}

\item{method}{a character specifying the distance method to be computed. Should be one of the following: "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis."}

\item{clustering}{a character specifying the clustering method to be used.}

\item{binary}{a boolean indicating whether or not to transform the data into a presence/absence data. Default is FALSE}

\item{label_colors}{a list of colors for each factor column in metaData. See \code{\link{plotColors}}. If NULL, default colors are used.}
}
\description{
This function assesses the repertoire dissimilarities by plotting a heatmap of a squared distance matrix computed between samples using a specific dissimilarity method.


It calculates a list of dissimilarity indices, each taking into account different parameters. The proposed methods include:

 The Jaccard similarity: a measure of similarity between sample sets defined as the size of the intersection divided by the size of the union of the sample sets.

 The Morisita-Horn similarity: a measure of similarity that tends to be over-sensitive to abundant species.

The function also performs a hierarchical clustering on the calculated distance scores.
}
\details{
Details on the calculated indices as well as the clustering methods can be found in the vegan package: https://www.rdocumentation.org/packages/vegan/versions/2.4-2/topics/vegdist
}
\examples{

data(RepSeqData)

plotDissimilarityMatrix(x = RepSeqData, level = "V", method = "euclidean")

plotDissimilarityMatrix(x = RepSeqData, level = "J", method = "euclidean")

plotDissimilarityMatrix(x = RepSeqData, level = "VJ", method = "euclidean")

plotDissimilarityMatrix(x = RepSeqData, level = "clone", method = "euclidean")

plotDissimilarityMatrix(x = RepSeqData, level = "clonotype", method = "euclidean")

plotDissimilarityMatrix(x = RepSeqData, level = "CDR3nt", method = "euclidean")

plotDissimilarityMatrix(x = RepSeqData, level = "CDR3aa", method = "euclidean")

plotDissimilarityMatrix(x = RepSeqData, level = "J", method = "manhattan")

}
