% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotGeneUsage}
\alias{plotGeneUsage}
\title{Compare V or J gene distributions}
\usage{
plotGeneUsage(
  x,
  level = c("V", "J"),
  scale = c("count", "frequency"),
  colorBy = NULL,
  facetBy = NULL,
  label_colors = NULL,
  show_stats = FALSE
)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken 
into account when calculating the gene usages. Should be one of "aaClone" or 
"ntClone".}

\item{scale}{a character specifying whether to plot the gene usage in "count" or "frequency".}

\item{colorBy}{a character indicating a column name in mData. Colors are thus attributed to the different groups within this column. The chosen column must be of class factor.}

\item{facetBy}{a vector of character indicating one or two column names in mData to apply a facet on.}

\item{label_colors}{a list of colors for each variable in ColorBy. See \code{\link{plotColors}}. If NULL, default colors are used.}

\item{show_stats}{whether to statistically compare groups}
}
\description{
This function compares the V or J gene usages between given groups.
}
\examples{

data(RepSeqData)

plotGeneUsage(x = RepSeqData, 
              level = "J", 
              scale = "count", 
              colorBy = "cell_subset", 
              show_stats=TRUE )

plotGeneUsage(x = RepSeqData,
             level = "V", 
             scale = "frequency",
             colorBy = "cell_subset", 
             facetBy="sex")

}
