% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotIndCountIntervals}
\alias{plotIndCountIntervals}
\title{Visualization of the clonal distribution per count interval in a single ample}
\usage{
plotIndCountIntervals(
  x,
  sampleName = NULL,
  level = c("clone", "clonotype", "CDR3nt", "CDR3aa"),
  fractions = c("count", "frequency", )
)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{sampleName}{a character specifying the sample_id to analyze. Default is NULL, which plots the first sample in the dataset.}

\item{level}{a character specifying the level of the repertoire to be taken into account when calculating the clonal distribution. Should be one of clone","clonotype", "CDR3nt" or "CDR3aa".}

\item{fractions}{whether fractions should be determined in count or frequency}
}
\description{
This function plots two histograms of the clonal distribution per a set of fixed count intervals in a given sample. Intervals are: 1, ]1, 10], ]10, 100], ]100, 1000], ]1000, 10000] and ]10000, Inf].

The upper plot calculates the proportion of each interval in the whole repertoire, whereas the lower plot shows the cumulative frequency of the sequences within each interval.

This allows a global view of the repertoire fraction contributing the most to the repertoire. For instance, top sequences belonging to the highest interval often constitute a low fraction of the whole repertoire but contribute more significantly in terms of cumulative frequency in view of their high occurrence.
}
\examples{

data(RepSeqData)
snames <- rownames(mData(RepSeqData))

plotIndCountIntervals(x = RepSeqData, level="CDR3aa",sampleName = snames[1])

plotIndCountIntervals(x = RepSeqData, level="clone", sampleName = NULL)

}
