% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotIndIntervals}
\alias{plotIndIntervals}
\title{Visualization of the clonal distribution per interval in a single sample}
\usage{
plotIndIntervals(
  x,
  sampleName = NULL,
  level = c("aaClone", "ntClone", "ntCDR3", "aaCDR3"),
  fractions = c("count", "frequency")
)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{sampleName}{a character specifying the sample_id to analyze. Default is NULL, which plots the first sample in the dataset.}

\item{level}{a character specifying the level of the repertoire to be taken into account when calculating the clonal distribution. Should be one of "aaClone","ntClone", "ntCDR3" or "aaCDR3".}

\item{fractions}{whether intervals should be determined in count or frequency}
}
\description{
This function plots two histograms of the clonal distribution per a set of fixed intervals in a given sample.

The upper plot calculates the proportion of each interval in the whole repertoire, whereas the lower plot shows the cumulative frequency of the sequences within each interval.

This allows a global view of the repertoire fraction contributing the most to the repertoire. For instance, top sequences belonging to the highest interval often constitute a low fraction of the whole repertoire but contribute more significantly in terms of cumulative frequency in view of their high occurrence.
}
\examples{

data(RepSeqData)

snames <- rownames(mData(RepSeqData))

plotIndIntervals(x = RepSeqData, level="aaCDR3", sampleName = snames[1],  fractions="count")

plotIndIntervals(x = RepSeqData, level="aaClone", sampleName = NULL,  fractions="frequency")

}
