% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotIndMap}
\alias{plotIndMap}
\title{Visualization of a repertoire structure through a circular treemap}
\usage{
plotIndMap(
  x,
  sampleName = NULL,
  prop = 0.01,
  level = c("aaClone", "ntClone", "V", "J", "VJ", "ntCDR3", "aaCDR3")
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{sampleName}{a character specifying the sample_id to analyze. Default is NULL, which plots the first sample in the dataset.}

\item{prop}{a numeric indicating the proportions of clones to be plotted. It ranges from 0 to 1.}

\item{level}{a character specifying the level of the repertoire on which the diversity should be estimated. Should be one of "aaClone","ntClone", "V", "J", "VJ", "ntCDR3" or "aaCDR3".}
}
\description{
This function plots all clones in a single repertoire as a circular treemap, showing the clone distributiion within the repertoire. Each circle represents a unique clone, and the circle size corresponds to the clone count.
}
\examples{

data(RepSeqData)

sname <- rownames(mData(RepSeqData))[1]

plotIndMap(x=RepSeqData, sampleName=sname, level="aaClone", prop=0.01)
}
