% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotPerturbationScore}
\alias{plotPerturbationScore}
\title{Visualization of the perturbation scores}
\usage{
plotPerturbationScore(
  x,
  ctrl.names = NULL,
  distance = c("manhattan", "euclidean", "canberra", "minkowski", "maximum"),
  p = 2,
  order = "cell_subset",
  label_colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{ctrl.names}{a vector of characters indicating the sample_ids to be used as controls.}

\item{distance}{a character specifying the distance method to be used in the calculation of the perturbation scores. Should be one of the following: "manhattan", "euclidean", "canberra", "minkowski" or "maximum".}

\item{p}{an integer indicating the power of the Minkowski distance. Default is 2.}

\item{order}{a character specifying the column name in mData to be used as reference to order the sample_ids in the heatmap, as no hierarchical clustering is performed in this analysis.}

\item{label_colors}{a list of colors for each variable in ColorBy. See \code{\link{plotColors}}. If NULL, default colors are used.}
}
\description{
plots the calculated perturbation scores using the \code{\link{perturbationScore}} function onto a heatmap.
}
\examples{

data(RepSeqData)
plotPerturbationScore(x = RepSeqData,
                      ctrl.names = c("tripod-30-813" ,"tripod-31-846"  ,"tripod-35-970"),
                      distance = "manhattan",
                      order= "cell_subset")
                      
}
\keyword{internal}
