% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotRarefaction}
\alias{plotRarefaction}
\title{Visualization of the rarefaction curves}
\usage{
plotRarefaction(x, colorBy = NULL, label_colors = NULL)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{colorBy}{a character indicating a column name in mData. Colors are thus attributed to the different groups within this column. The chosen column must be of class factor.}

\item{label_colors}{a list of colors for each variable in ColorBy. See \code{\link{plotColors}}. If NULL, default colors are used.}
}
\value{
none
}
\description{
This function plots the rarefaction curve for each sample within the dataset.

Rarefaction is a measure of species richness. The curves plot the number of clones against the number of sequences in a sample, each being obtained by randomly re-sampling a number of sequences multiple times and representing the mean number of found clones.
}
\examples{

data(RepSeqData)

plotRarefaction(x = RepSeqData, colorBy = "sex")

plotRarefaction(x = RepSeqData, colorBy = "cell_subset")


}
