% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotRenyiIndex}
\alias{plotRenyiIndex}
\title{Visualization of the Renyi index}
\usage{
plotRenyiIndex(
  x,
  alpha = c(0, 0.25, 0.5, 1, 2, 4, 8, 16, 32, 64, Inf),
  level = c("clone", "clonotype", "V", "J", "VJ", "CDR3nt", "CDR3aa"),
  colorBy = NULL,
  grouped = FALSE,
  label_colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{alpha}{a numerical vector specifying the alpha values to compute. If not specified, the following values are estimated: c(0, 0.25, 0.5, 1, 2, 4, 8, 16, 32, 64, Inf).}

\item{level}{a character specifying the level of the repertoire to be taken into account when calculating VJ usages. Should be one of clone","clonotype", "V", "J", "VJ", "CDR3nt" or "CDR3aa".}

\item{colorBy}{a character indicating a column name in mData. Colors are thus attributed to the different groups within this column. The chosen column must be of class factor.}

\item{grouped}{a boolean indicating whether or not the mean and se of samples belonging to the same experimental group specified in the ColorBy parameter should be calculated. Default is FALSE. The area under the curve (AUC) is calculated for each sample and compared between chosen groups using a Wilcoxon test with Holm's correction.}

\item{label_colors}{a list of colors for each variable in ColorBy. See \code{\link{plotColors}}. If NULL, default colors are used.}
}
\description{
This function plots the Renyi values at any repertoire level for all the samples in the dataset.
The alpha values for which the Renyi is to be estimated can be personalized, thus allowing to focus on certain indices such as the Shannon index for alpha=1 or the Simpson index for alpha=2.
For instance, the most frequent sequence is attributed a rank of 1, and its relative abundance is plotted on the y-axis.
}
\details{
The Renyi index is a generalization of the Shannon index. It represents the distribution of clonal expansions
as a function of the parameter alpha. At alpha=0, it equally considers all species including the rare ones,
whereas it up-weighs the abundant species with an increasing value of alpha. Alpha =1 is an approximation of the Shannon index; 
alpha = 2 corresponds to the Simpson index and alpha=Inf corresponds to the Berger-Parker index. The latter highlights the highest clonal expansion in a repertoire.
}
\examples{

data(RepSeqData)
plotRenyiIndex(x = RepSeqData, level = "V", colorBy = "sex")

plotRenyiIndex(x = RepSeqData, level = "J", colorBy = "sex", grouped=TRUE, groupBy="sex")

plotRenyiIndex(x = RepSeqData, level = "J", colorBy = "sex", grouped=TRUE, groupBy=c("sex", "cell_subset"))

}
