% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotScatter}
\alias{plotScatter}
\title{Visualization of the inter-repertoire sharing on a scatter plot}
\usage{
plotScatter(
  x,
  sampleNames = NULL,
  level = c("aaClone", "ntClone", "V", "J", "VJ", "ntCDR3", "aaCDR3"),
  scale = c("frequency", "count", "log"),
  shiny = FALSE
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{sampleNames}{a vector of character indicating the two sample_ids of the repertoires to be analyzed.}

\item{level}{a character specifying the level of the repertoire to be analyzes. Should be one of "aaClone", "ntClone", "V", "J", "VJ", "ntCDR3" or "aaCDR3".}

\item{scale}{a character specifying whether to plot in "count" , "frequency". or "log'}

\item{shiny}{default is FALSE. whether to plot the shiny compatible version}
}
\description{
This function computes a correlation between a pair of repertoires at any repertoire level.
It allows a comparative analysis of the occurrence of a given repertoire level between a pair of samples, which goes beyond the simple calculation of the sharing degree.
}
\examples{

data(RepSeqData)
plotScatter(x = RepSeqData,
            level = "V",
            sampleNames = c("tripod-30-813", "tripod-30-815", "tripod-31-846"),
            scale = "log")
            
plotScatter(x = RepSeqData,
            level = "aaClone",
            sampleNames = c("tripod-30-813", "tripod-30-815"),
            scale = "frequency")


}
