% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotSpectratyping}
\alias{plotSpectratyping}
\title{Visualization of CDR3 spectratyping}
\usage{
plotSpectratyping(
  x,
  sampleName = NULL,
  scale = c("count", "frequency"),
  prop = 0
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{sampleName}{a character specifying the sample_id to analyze. Default is NULL, which plots the first sample in the dataset.}

\item{scale}{a character specifying whether to plot the length usage in "count" or "frequency".}

\item{prop}{a numeric indicating the proportions of top V genes to be plotted. It ranges from 0 to 1.}
}
\description{
CDR3 spectratyping allows large-scale analysis of the repertoire diversity based on the CDR3 amino acid length usage.

With this technique, a polyclonal repertoire is represented by an eight-peak bell-shaped profile, that is disturbed following an immune response.

This function plots a histogram of the CDR3 length distribution in a single sample.
It is also possible to add the information on the V gene usage per CDR3 length.
To do so, the proportion of the most used V genes within the repertoire can be specified with the prop parameter, thus showing their relative proportions within each CDR3 length.

Moreover, this function integrates a statistical method that can be applied on the spectratypes in order to compare peak distributions across different groups of samples called ISEApeaks (Colette and Six, 2002).
The strategy can be applied using the \code{\link{perturbationScore}} function.
}
\examples{

data(RepSeqData)
snames <- rownames(mData(RepSeqData))

plotSpectratyping(x = RepSeqData, sampleName = NULL, scale = "count", prop=1)

plotSpectratyping(x = RepSeqData, sampleName = NULL, scale = "frequency", prop=0)

}
\seealso{
\code{\link{perturbationScore}}
}
\keyword{internal}
