% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotSpectratypingV}
\alias{plotSpectratypingV}
\title{Visualization of the CDR3 spectratyping per V gene}
\usage{
plotSpectratypingV(
  x,
  sampleName = NULL,
  scale = c("count", "frequency"),
  prop = 0
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{sampleName}{a character specifying the sample_id to analyze. Default is NULL, which plots the first sample in the dataset.}

\item{scale}{a character specifying whether to plot the length usage in "count" or "frequency".}

\item{prop}{a numeric indicating the proportions of top V genes to be plotted. It ranges from 0 to 1.}
}
\description{
CDR3 spectratyping allows large-scale analysis of the repertoire diversity based on the CDR3 amino acid length usage.
With this technique, a polyclonal repertoire is represented by an eight-peak bell-shaped profile that is disturbed following an immune response.

This function plots a histogram of the CDR3 length distribution per V gene in a single sample.
}
\examples{

data(RepSeqData)

snames <- rownames(mData(RepSeqData))

plotSpectratypingV(x = RepSeqData, sampleName = snames[1], scale = "count",prop=0.05)

plotSpectratypingV(x = RepSeqData, sampleName = snames[1], scale = "frequency")


}
