% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotStatistics}
\alias{plotStatistics}
\title{Visualization of basic statistics}
\usage{
plotStatistics(
  x,
  stat = c("nSequences", "aaClone", "ntClone", "V", "J", "VJ", "aaCDR3", "ntCDR3"),
  grouped = FALSE,
  colorBy = NULL,
  facetBy = NULL,
  label_colors = NULL,
  show_stats = FALSE
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{stat}{a character specifying the statistic to plot. Should one of the statistics in the metaData slot.}

\item{grouped}{a character indicating one or multiple groups to be compared. A Wilcoxon test is thus performed and adjusted p-values using the Holm method are shown. Colors are attributed to the different groups within the first column, and a facet is applied on the second column. If not specified, no statistical tests will be performed, and calculated values for each sample_id will be represented.}

\item{colorBy}{a character indicating a column name in mData. Colors are thus attributed to the different groups within this column. The chosen column must be of class factor.}

\item{facetBy}{a vector of character indicating one or two column names in mData to apply a facet on.}

\item{label_colors}{a list of colors for each variable in groupBy See \code{\link{plotColors}}. If NULL, default colors are used.}

\item{show_stats}{whether to statistically compare groups}
}
\description{
This function plots the statistics in the mData slot, calculated for each sample during the building of the RepSeqExperiment object.

These statistics include:

- nSequences: the total number of sequences in a sample

- ntCDR3: the number of unique nucleotide CDR3s

- aaCDR3: the number of unique amino acid CDR3s

- V: the total number of V genes expressed in each sample

- J: the total number of J genes

- VJ: the total number of V-J gene combinations

- aaClone: the number of unique aaClones

- ntClone: the number of unique ntClones

They can be compared between groups of samples or simply plotted for each sample.
}
\examples{

data(RepSeqData)

plotStatistics(x = RepSeqData, stat = "V", colorBy = "sex")

plotStatistics(x = RepSeqData, colorBy = "cell_subset", facetBy="sex", stat = "aaClone", grouped=TRUE)

}
