% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotVenn}
\alias{plotVenn}
\title{Visualization of the inter-repertoire sharing on an Venn diagram}
\usage{
plotVenn(
  x,
  level = c("aaClone", "ntClone", "V", "J", "VJ", "ntCDR3", "aaCDR3"),
  sampleNames = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire on which the diversity should be estimated. Should be one of "aaClone","ntClone", "V", "J", "VJ", "ntCDR3" or "aaCDR3".}

\item{sampleNames}{a vector of character indicating the sample_ids of the repertoires to be analyzed. If not specified, the first three samples in the dataset are analyzed.}
}
\description{
The repertoire sharing at any level evaluates the degree of convergence between repertoires and experimental conditions.

This function plots the number of shared sequences, at any repertoire level, between samples belonging for instance, to a same experimental group.
No limitations in the number of sample_ids is imposed, however for visualization purposes, a maximum of 5 sample_ids should be ideally specified.
}
\examples{

data(RepSeqData)

snames <- rownames(mData(RepSeqData))[1:4]

plotVenn(x = RepSeqData, level = "V", sampleNames = snames)

plotVenn(x = RepSeqData, level = "aaCDR3", sampleNames = NULL)

}
