% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{readAIRR}
\alias{readAIRR}
\title{Read a single alignment file}
\usage{
readAIRR(
  path,
  fileFormat = c("MiXCR", "immunoseq", "MiAIRR"),
  chain = c("TRA", "TRB", "TRG", "TRD", "IGH", "IGK", "IGL"),
  keep.ambiguous = FALSE,
  keep.unproductive = FALSE,
  aa.th = NULL,
  outFiltered = FALSE
)
}
\arguments{
\item{path}{a path to an output file of the supported aligners.}

\item{fileFormat}{a character, the name of the aligner. Should be one of "immunoseq", "MiXCR" or "MiAIRR". Default is MiXCR.}

\item{chain}{a character, indicates which TCR chain to import: \code{A} for the alpha chain and \code{B} for the beta chain. Default is \code{A}.}

\item{keep.ambiguous}{a boolean, whether ambiguous sequences (containing a STOP codon) should be kept. Default is \code{FALSE}.}

\item{keep.unproductive}{a boolean, whether unproductive sequences (Euclidean dividion of aa length by 3 > 0) should be kept. Default is \code{FALSE}.}

\item{aa.th}{an interger, indicates the maximum number of amino acids deviating from the mean length to tolerate. Default is NULL.}

\item{outFiltered}{write in the directory path a data frame containing the columns of the input clonotype table and the filterd out reads as rows. Default path is getwd().}
}
\value{
a filtered clonotype table.
}
\description{
parse and filter clonotype tables.
}
\details{
this function is a wrapper of parse functions & filterClones
}
\examples{

l <- list.files(system.file(file.path('extdata/MiAIRR'),
                     package = 'AnalyzAIRR'),
                     full.names = TRUE)


dataset <- readAIRR(path = l[1],
                          fileFormat = "MiAIRR",
                          chain = "TRA",
                          keep.ambiguous = FALSE,
                          keep.unproductive = FALSE,
                         outFiltered = FALSE)

}
\keyword{internal}
