% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{readInFormat}
\alias{readInFormat}
\title{read in format}
\usage{
readInFormat(
  path,
  chain = c("TRA", "TRB", "TRG", "TRD", "IGH", "IGK", "IGL"),
  keep.ambiguous = FALSE,
  keep.unproductive = FALSE,
  aa.th = NULL,
  outFiltered = FALSE
)
}
\arguments{
\item{path}{full path to the aligned file. Files can be loaded as gzipped.}

\item{chain}{character, the TCR chain to be analyzed. One of \code{A} or \code{B}. Default value is \code{A}.}

\item{keep.ambiguous}{a boolean, whether ambiguous sequences containing an "N" in the CDR3 nucleotide should be left in. Default is \code{FALSE}.}

\item{keep.unproductive}{a boolean, whether unproductive sequences should be left in. Default is \code{FALSE}.
Unproductive sequences include:

- out-of-frame sequences: sequences with frame shifts based on the number of nucleotides in the ntCDR3 column

- sequences containing stop codons: aaCDR3 sequences with a "*" or "~" symbols.should be kept. Default is \code{FALSE}.}

\item{aa.th}{an interger, indicates the maximum number of amino acids deviating from the mean length to tolerate. Default is 8. In this case, all amino acid CDR3s sequences with length falling outside the range of mean+- 8 are filtered out.}

\item{outFiltered}{write in the directory path a data frame containing the filtered out reads. Default path is getwd().}
}
\value{
a data.table of 9 columns: \code{sample_id} name, \code{V} V gene, \code{J} J gene,\code{VJ} V-J gene combinations, \code{aaCDR3} amino acid CDR3 sequence, \code{ntCDR3} nucleotide CDR3 sequence, \code{ntClone} full clone sequence in nucleotide, \code{aaClone} full clone sequence in amino acid, \code{count} clone count. Clones are eliminated if a STOP codon (*) is detected in aaCDR3 chain, if the ntCDR3 length is not divisible by 3 or if the ntCDR3 sequence is ambiguous (contains a "N" base).
}
\description{
alternative
}
\keyword{internal}
