% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{readSingleCellSet}
\alias{readSingleCellSet}
\title{A method for the building of a RepSeqExperiment object using single cell data}
\usage{
readSingleCellSet(
  fileList,
  sampleinfo = NULL,
  filter.singletons = FALSE,
  cores = 1L
)
}
\arguments{
\item{fileList}{a path to the acell ranger filtered contigs file}

\item{sampleinfo}{a path to the acell ranger filtered contigs file}

\item{filter.singletons}{a path to the acell ranger filtered contigs file}

\item{cores}{a path to the acell ranger filtered contigs file}
}
\value{
an object of class \code{\link{RepSeqExperiment}}. See \code{\link{RepSeqExperiment-class}} for more details.
}
\description{
This function
}
\keyword{internal}
