% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepSeqExperiment.R
\docType{class}
\name{RepSeqExperiment-class}
\alias{RepSeqExperiment-class}
\alias{RepSeqExperiment}
\alias{assay,RepSeqExperiment-method}
\alias{assay}
\alias{assay<-,RepSeqExperiment-method}
\alias{assay<-}
\alias{oData,RepSeqExperiment-method}
\alias{oData}
\alias{oData<-,RepSeqExperiment-method}
\alias{oData<-}
\alias{mData,RepSeqExperiment-method}
\alias{mData}
\alias{mData<-,RepSeqExperiment-method}
\alias{mData<-}
\alias{History,RepSeqExperiment-method}
\alias{History}
\alias{History<-,RepSeqExperiment-method}
\alias{History<-}
\alias{show,RepSeqExperiment-method}
\alias{names,RepSeqExperiment-method}
\alias{names}
\alias{names<-,RepSeqExperiment,ANY-method}
\alias{names<-}
\alias{names<-,RepSeqExperiment-method}
\alias{[,RepSeqExperiment,ANY,ANY,ANY-method}
\alias{[,RepSeqExperiment-method}
\alias{is.RepSeqExperiment}
\title{The class RepSeqExperiment}
\usage{
assay(object)

assay(object, i, j) <- value

oData(object)

oData(object) <- value

mData(object)

mData(object) <- value

History(object)

History(object) <- value

\S4method{assay}{RepSeqExperiment}(object)

\S4method{assay}{RepSeqExperiment}(object, i, j) <- value

\S4method{oData}{RepSeqExperiment}(object)

\S4method{oData}{RepSeqExperiment}(object) <- value

\S4method{mData}{RepSeqExperiment}(object)

\S4method{mData}{RepSeqExperiment}(object) <- value

\S4method{History}{RepSeqExperiment}(object)

\S4method{History}{RepSeqExperiment}(object) <- value

\S4method{show}{RepSeqExperiment}(object)

\S4method{names}{RepSeqExperiment}(x)

\S4method{names}{RepSeqExperiment,ANY}(x) <- value

\S4method{[}{RepSeqExperiment,ANY,ANY,ANY}(x, i, j, drop)

is.RepSeqExperiment(x)
}
\arguments{
\item{object}{a RepSeqExperiment object}

\item{i}{indice(s) of clonotype(s) to extract}

\item{j}{indice(s) of sample(s) to extract}

\item{value}{either numeric, character or data frame}

\item{x}{an object}

\item{drop}{If \code{TRUE} the result is coerced to the lowest possible dimension}
}
\value{
an object of class [\code{\linkS4class{RepSeqExperiment}}]
}
\description{
An S4 class object enclosing the adaptive immune repertoire data, and which is used in all the analytical metrics proposed by the AnalyzAIRR package. The RepSeqExperiment object is composed of 4 slots, each containing different information.

\code{[]} Extract parts of the \code{\linkS4class{RepSeqExperiment}} object
}
\section{Slots}{

\describe{
\item{\code{assayData}}{a data.table binding all clonotype tables and containing the following columns:

 - sample_id: sample names

 - V: Variable gene name

 - J: Joining gene name

 - CDR3aa: CDR3 amino acid sequence

 - CDR3nt: CDR3 nucleotide sequence

 - clone: Full clonotype sequence including the V gene, the amino acid CDR3 sequence and the J gene

 - VJ: V-J gene combination

 - count: the occurrence of the clonotype, i.e the clone sequence}

\item{\code{metaData}}{a data frame containing sample information specified during the building of the \code{\linkS4class{RepSeqExperiment}} object. Each row represents a sample, and columns the possible information or group that can be attributed to the samples such as the cell population, the donor's age, sex, etc...}

\item{\code{otherData}}{a list of meta data containing any other information than the user does not wish to include in the analysis.}

\item{\code{History}}{a data frame registering all operations performed on \code{\linkS4class{RepSeqExperiment}} object, such as the filtering functions and the normalization.}
}}

