% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{theme_RepSeq}
\alias{theme_RepSeq}
\title{A theme for the visualization plots}
\usage{
theme_RepSeq()
}
\description{
A unique plot theme used in all the visualization plots.
As this function is generated using the ggplot2 package, users can easily
customize all of the generated plots using ggplot2 arguments.
}
