% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzAIRR.R
\name{AnalyzAIRR}
\alias{AnalyzAIRR}
\title{AnalyzAIRR - AIRR-compliant R package for the analysis of bulk Ig/TCR repertoire datasets}
\description{
AnalyzAIRR is a AIRR-compliant R package developed to analyze bulk Ig/TCR repertoire datasets.
}
\details{
AnalyzAIRR allows a general data exploration to evaluate the homogeneity 
within defined groups, identify outliers and filter them out. AnalyzAIRR also proposes
a set of diversity measures and statistical metrics applicable at any level of granularity. Thus, single-sample repertoire explorations or in-depth cross-comparisons of AIRR datasets 
can be conducted leading to ready-to-publish visualization graphics. 
AnalyzAIRR is complemented with a guided workflow to help users in their analytical strategy, and a Shiny web application making it user-friendly for biologists with little or no background in bioinformatics.
}
\seealso{
Useful links:
\itemize{
 \item \href{https://github.com/i3-unit/AnalyzAIRR}{https://github.com/i3-unit/AnalyzAIRR}
 \item \href{https://github.com/vanessajmh/Shiny-AnalyzAIRR}{https://github.com/vanessajmh/Shiny-AnalyzAIRR}
}
}
\author{
\itemize{
\item Vanessa Mhanna
\item Gabriel Pires
\item Grégoire Bohl
\item Karim El Soufi
\item Nicolas Tchitchek
\item David Klatzmann
\item Adrien Six
\item Hang-Phuong Pham
\item Encarnita Mariotti-Ferrandiz 
}
}
