% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFunctions.R
\name{diffExpGroup}
\alias{diffExpGroup}
\title{Differential expression analysis}
\usage{
diffExpGroup(
  x,
  colGrp,
  level = c("aaClone", "ntClone", "V", "J", "VJ", "ntCDR3", "aaCDR3"),
  group
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{colGrp}{a vector of character specifying the column names in the mData slot corresponding to the experimental condition to be analyzed.}

\item{level}{a character specifying the level of the repertoire to be compared. Should be one of "aaClone","ntClone", "V", "J", "VJ", "ntCDR3" or "aaCDR3".}

\item{group}{a vector of character indicating the column name in the mData slot, as well as the two groups to be compared.}
}
\value{
a data.frame with 6 columns: the repertoire level in rownames, and the baseMean, log2FoldChange, lfcSE, stat, pvalue and padj in columns.
The table is ordered by adjusted p-values.
}
\description{
This function identifies differentially expressed repertoire levels between groups of samples.
}
\details{
This function uses the \href{https://www.rdocumentation.org/packages/DESeq2/versions/1.12.3/topics/DESeq}{DESeq2} package.
Briefly, it estimates the size factors using the poscounts method which deals with zero counts.
It then performs a default analysis by estimating the dispersion using a local regression of log dispersions over log base mean.
Finally, a generalized linear model is fitted using a Negative Binomial distribution and Wald statistics.
}
\examples{
if (requireNamespace("DESeq2", quietly = TRUE)) {
dds1 <- diffExpGroup(x = RepSeqData,
                    colGrp = "cell_subset" ,
                    level = "V",
                    group = c("cell_subset", "amTreg", "nTreg"))
}

}
\references{
Hill, M.O. (1973). Diversity and evenness: a unifying notation and its consequences. Ecology 54, 427–473.
Kindt, R., Van Damme, P., Simons, A.J. (2006). Tree diversity in western Kenya: using profiles to characterise richness and evenness. Biodiversity and Conservation 15, 1253–1270.
}
