% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{filterSequence}
\alias{filterSequence}
\title{Exclude a sequence from a RepSeqExperiment object.}
\usage{
filterSequence(
  x,
  level = c("clone", "clonotype", "CDR3aa", "CDR3nt"),
  name,
  group = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken into account. Should be one of "clone","clonotype", "CDR3nt" or "CDR3aa".}

\item{name}{a vector of character specifying the name(s) of the sequence(s) to filter out from the RepSeqExperiment object.}

\item{group}{a vector of character indicating the group column name in the mData slot and one experimental group within this column.}
}
\value{
a RepSeqExperiment object.
}
\description{
This function allows the dropping of one or several sequences from a RepSeqExperiment object in all samples or a specified group of samples.
}
\examples{

data(RepSeqData)

RepSeqData<- filterSequence(x = RepSeqData,
                            level="clone",
                            name="TRAV11 CVVGDRGSALGRLHF TRAJ18",
                            group=c("cell_subset","Teff"))

}
