% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getPrivate}
\alias{getPrivate}
\title{Extract private sequences}
\usage{
getPrivate(
  x,
  level = c("clone", "clonotype", "CDR3aa", "CDR3nt"),
  singletons = FALSE
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken into account. Should be one of clone","clonotype", "CDR3nt" or "CDR3aa".}

\item{singletons}{a boolean indicating whether or not private sequences with a count of 1 should be extracted. Default is FALSE.}
}
\value{
an object of class \code{\linkS4class{RepSeqExperiment}} composed exclusively of private sequences.
}
\description{
This function extract private sequences within the whole dataset, i.e. sequences found exclusively in one sample.
}
\examples{

data(RepSeqData)

privateclones <- getPrivate(RepSeqData, level = "clone", singletons = TRUE)

privateclonotypes <- getPrivate(RepSeqData, level = "clonotype", singletons = FALSE)

privateCDR3nt <- getPrivate(RepSeqData, level = "CDR3nt", singletons = TRUE)

privateCDR3aa <- getPrivate(RepSeqData, level = "CDR3aa", singletons = FALSE)

}
