% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{getTopSequences}
\alias{getTopSequences}
\title{Extract the most frequent sequences}
\usage{
getTopSequences(
  x,
  level = c("clone", "clonotype", "CDR3aa", "CDR3nt"),
  group = NULL,
  prop = 0
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken into account. Should be one of clone","clonotype", "CDR3nt" or "CDR3aa".}

\item{group}{character, column name in sData. Must be of class factor. Default is NULL, the threshold is calculated across all the samples within the dataset.}

\item{prop}{a numeric between 0 and 1 indicating the proportion of top sequences to extract.}
}
\value{
an object of class \code{\linkS4class{RepSeqExperiment}}
}
\description{
This function extracts the top n sequences within all samples or a group of samples.
}
\examples{

data(RepSeqData)
topClones <- getTopSequences(RepSeqData,
                          level = "clone",
                          group = c("cell_subset", "amTreg"))


}
