% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotGeneUsage}
\alias{plotGeneUsage}
\title{Compare V or J gene distributions}
\usage{
plotGeneUsage(
  x,
  level = c("V", "J"),
  scale = c("count", "frequency"),
  groupBy = NULL,
  label_colors = NULL
)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken into account when calculating the gene usages. Should be one of clone" or "clonotype".}

\item{scale}{a character specifying whether to plot the gene usage in "count" or "frequency".}
}
\description{
This function compares the V or J gene usages between given groups.
}
\examples{

data(RepSeqData)

plotGeneUsage(x = RepSeqData, level = "J", scale = "count", groupBy="cell_subset")

plotGeneUsage(x = RepSeqData, level = "V",  scale = "frequency", groupBy="sex")

}
