% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotIndGeneUsage}
\alias{plotIndGeneUsage}
\title{Visualization of the V or J gene usage}
\usage{
plotIndGeneUsage(
  x,
  sampleName = NULL,
  level = c("V", "J"),
  scale = c("count", "frequency")
)
}
\arguments{
\item{x}{an object of class  \code{\linkS4class{RepSeqExperiment}}}

\item{sampleName}{a character specifying the sample_id to analyze. Default is NULL, which plots the first sample in the dataset.}

\item{level}{a character specifying the level of the repertoire to be taken into account when calculating the gene usages. Should be one of "aaClone" or "ntClone".}

\item{scale}{a character specifying whether to plot the gene usage in "count" or "frequency".}
}
\description{
This function plots the V or J gene usages for a given sample.
}
\examples{

data(RepSeqData)
snames <- rownames(mData(RepSeqData))

plotIndGeneUsage(x = RepSeqData, level = "V", sampleName = snames[1], scale = "count")

plotIndGeneUsage(x = RepSeqData, level = "J", sampleName = NULL, scale = "frequency")
}
