% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotRankDistrib}
\alias{plotRankDistrib}
\title{Visualization of the clonal distribution as a function of the rank}
\usage{
plotRankDistrib(
  x,
  level = c("aaClone", "ntClone", "ntCDR3", "aaCDR3"),
  scale = c("count", "frequency"),
  ranks = NULL,
  grouped = FALSE,
  colorBy = NULL,
  facetBy = NULL,
  label_colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken into account to plot the clonal distribution. Should be one of "aaClone", "ntClone", "ntCDR3" or "aaCDR3".}

\item{scale}{a character specifying whether to plot the clonal abundance in "count" or "frequency".}

\item{ranks}{an integer specifying the number of top ranks to be plotted. Default is NULL, which plots all ranks.}

\item{grouped}{a boolean indicating whether or not the mean and se of samples belonging to the same experimental group specified in the ColorBy parameter should be computed. Grouping is performed on the chosen groupe in colorBy. Default is FALSE.}

\item{colorBy}{a character indicating one column name in mData. Colors are thus attributed to the different groups within this column. The chosen column must be of class factor.}

\item{facetBy}{a vector of character indicating one or two column names in mData to apply a facet on.}

\item{label_colors}{a list of colors for each variable in ColorBy. See \code{\link{plotColors}}. If NULL, default colors are used.}
}
\description{
This function plots the clonal distribution, at any repertoire level, as a function of the occurrence rank within each sample.

For instance, the most frequent sequence is attributed a rank of 1, and its relative abundance is plotted on the y-axis.
}
\examples{

data("RepSeqData")

plotRankDistrib(x = RepSeqData, level="ntClone", colorBy = "sample_id", 
                facetBy="sex", scale = "count")

plotRankDistrib(x = RepSeqData, level="aaClone", colorBy = "cell_subset", 
                grouped=TRUE, scale = "frequency")

plotRankDistrib(x = RepSeqData,level="ntCDR3", colorBy = "cell_subset", facetBy="sex", 
                grouped=TRUE,  scale = "frequency")

}
