% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotRankDistrib}
\alias{plotRankDistrib}
\title{Visualization of the clonal distribution as a function of the rank}
\usage{
plotRankDistrib(
  x,
  level = c("clone", "clonotype", "CDR3nt", "CDR3aa"),
  scale = c("count", "frequency"),
  colorBy = NULL,
  grouped = FALSE,
  label_colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{level}{a character specifying the level of the repertoire to be taken into account to plot the clonal distribution. Should be one of "clone", "clonotype", "CDR3nt" or "CDR3aa".}

\item{scale}{a character specifying whether to plot the clonal abundance in "count" or "frequency".}

\item{colorBy}{a character indicating a column name in mData. Colors are thus attributed to the different groups within this column. The chosen column must be of class factor.}

\item{grouped}{a boolean indicating whether or not the mean and se of samples belonging to the same experimental group specified in the ColorBy parameter should be computed. Default is FALSE. The area under the curve (AUC) is calculated for each sample and compared between chosen groups using a Wilcoxon test with Holm's correction.}

\item{label_colors}{a list of colors for each variable in ColorBy. See \code{\link{plotColors}}. If NULL, default colors are used.}
}
\description{
This function plots the clonal distribution, at any repertoire level, as a function of the occurrence rank within each sample.

For instance, the most frequent sequence is attributed a rank of 1, and its relative abundance is plotted on the y-axis.
}
\examples{

data("RepSeqData")

plotRankDistrib(x = RepSeqData, level="clone",colorBy = "sex", scale = "count")

plotRankDistrib(x = RepSeqData,level="clonotype", colorBy = "cell_subset", scale = "frequency")


}
