% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotStatistics}
\alias{plotStatistics}
\title{Visualization of basic statistics}
\usage{
plotStatistics(
  x,
  stat = c("nSequences", "clone", "clonotype", "V", "J", "VJ", "CDR3aa", "CDR3nt"),
  groupBy = NULL,
  label_colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{stat}{a character specifying the statistic to plot. Should one of the statistics in the metaData slot.}

\item{groupBy}{a character indicating the groups to be compared. A Wilcoxon test is thus performed and adjusted p-values using the Holm method are shown. If not specified, no statistical tests will be performed, and calculated values for each sample_id will be represented.}

\item{label_colors}{a list of colors for each variable in groupBy See \code{\link{plotColors}}. If NULL, default colors are used.}
}
\description{
This function plots the statistics in the mData slot, calculated for each sample during the building of the RepSeqExperiment object.

These statistics include:

- nSequences: the total number of sequences in a sample

- CDR3nt: the number of unique nucleotide CDR3s

- CDR3aa: the number of unique amino acid CDR3s

- V: the total number of V genes expressed in each sample

- J: the total number of J genes

- VJ: the total number of V-J gene combinations

- clone: the number of unique clones

- clonotype: the number of unique clonotypes

They can be compared between groups of samples or simply plotted for each sample.
}
\examples{

data(RepSeqData)

plotStatistics(x = RepSeqData, groupBy = "sex", stat = "V")

plotStatistics(x = RepSeqData, groupBy = "cell_subset", stat = "J")

plotStatistics(x = RepSeqData, groupBy = "sex", stat = "VJ")

plotStatistics(x = RepSeqData, groupBy = "cell_subset", stat = "clone")

plotStatistics(x = RepSeqData, groupBy = "sex", stat = "clonotype")

plotStatistics(x = RepSeqData, groupBy = "cell_subset", stat = "CDR3nt")

plotStatistics(x = RepSeqData, groupBy = "sex", stat = "nSequences")

}
