% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotStatistics}
\alias{plotStatistics}
\title{Visualization of basic statistics}
\usage{
plotStatistics(
  x,
  stat = c("nSequences", "clone", "clonotype", "V", "J", "VJ", "CDR3aa", "CDR3nt"),
  groupBy = NULL,
  label_colors = NULL,
  show_stats = TRUE
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{stat}{a character specifying the statistic to plot. Should one of the statistics in the metaData slot.}

\item{groupBy}{a character indicating one or multiple groups to be compared. A Wilcoxon test is thus performed and adjusted p-values using the Holm method are shown. Colors are attributed to the different groups within the first column, and a facet is applied on the second column. If not specified, no statistical tests will be performed, and calculated values for each sample_id will be represented.}

\item{label_colors}{a list of colors for each variable in groupBy See \code{\link{plotColors}}. If NULL, default colors are used.}

\item{show_stats}{whether to statistically compare groups}
}
\description{
This function plots the statistics in the mData slot, calculated for each sample during the building of the RepSeqExperiment object.

These statistics include:

- nSequences: the total number of sequences in a sample

- CDR3nt: the number of unique nucleotide CDR3s

- CDR3aa: the number of unique amino acid CDR3s

- V: the total number of V genes expressed in each sample

- J: the total number of J genes

- VJ: the total number of V-J gene combinations

- clone: the number of unique clones

- clonotype: the number of unique clonotypes

They can be compared between groups of samples or simply plotted for each sample.
}
\examples{

data(RepSeqData)

plotStatistics(x = RepSeqData, groupBy = "sex", stat = "V")

plotStatistics(x = RepSeqData, groupBy = c("cell_subset", "sex"), stat = "V")

}
