% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotVJusage}
\alias{plotVJusage}
\title{Visualization of the VJ combination usage}
\usage{
plotVJusage(
  x,
  sampleName = NULL,
  scale = c("count", "frequency"),
  level = c("clone", "clonotype"),
  prop = 1
)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{RepSeqExperiment}}}

\item{sampleName}{a character specifying the sample_id to analyze. Default is NULL, which plots the first sample in the dataset.}

\item{scale}{a character specifying whether to plot the VJ usage in "count" or "frequency".}

\item{level}{a character specifying the level of the repertoire to be taken into account when calculate VJ usages. Should be one of "clone" or "clonotype".}

\item{prop}{a numeric indicating the proportions of top VJ combinations to be plotted. It ranges from 0 to 1.}
}
\description{
This function calculates the count or frequency of each combination by taking into account the weight of the chosen repertoire level: either "clone" or "clonotype".

It offers two types of visualization of the calculated VJ usage in a given sample.

- For prop=1, a heatmap with all the VJ combinations is plotted, with the V genes as columns and the J genes as rows.

- For prop!=1, a circos plot is drawn showing the top proportions of VJ combinations specified in the prop parameter.
}
\examples{

data(RepSeqData)
snames <- rownames(mData(RepSeqData))

plotVJusage(x = RepSeqData, sampleName = NULL, scale = "count", level="clone", prop=1)

plotVJusage(x = RepSeqData, sampleName = NULL, scale = "count", level="clonotype", prop=0.1)

plotVJusage(x = RepSeqData, sampleName = NULL, scale = "frequency", level="clone", prop=0)

plotVJusage(x = RepSeqData, sampleName = NULL, scale = "frequency", level="clonotype")

plotVJusage(x = RepSeqData, sampleName = snames[1], scale = "frequency", level="clone")

plotVJusage(x = RepSeqData, sampleName = snames[1], scale = "frequency", level="clonotype", prop=0)

plotVJusage(x = RepSeqData, sampleName = snames[2], scale = "count", level="clone", prop=0.05)

plotVJusage(x = RepSeqData, sampleName = snames[2], scale = "frequency", level="clonotype")


}
