% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{RepSeqExp}
\alias{RepSeqExp}
\title{An alternative method for the building of a RepSeqExperiment object}
\usage{
RepSeqExp(clonotypetab, sampleinfo = NULL)
}
\arguments{
\item{clonotypetab}{a single clonotype table containnig the previously cited columns}

\item{sampleinfo}{a data frame containing:

- a column with the sample names. Names should match the base names of the corresponding files and their order. This column should be assigned as row.names when the metadata file is loaded. See the example below.

- any additional columns with relevant information for the subsequent analyses. Group columns must be transformed into factors after loading. See the example below.

No specific column names are required.}
}
\value{
an object of class \code{RepSeqExperiment} that is  used in all the analytical metrics proposed by the AnalyzAIRR package. See \code{\link{RepSeqExperiment-class}} for more details.
}
\description{
This function can be used to build a RepSeqExperiment object using aligned files that were not produced by a AnalyzAIRR-supported aligning tool. The loaded clonotype table must however contain the following required columns:

 - sample_id: sample names

 - V: Variable gene name

 - J: Joining gene name

 - CDR3aa: CDR3 amino acid sequence

 - CDR3nt: CDR3 nucleotide sequence

 - clone: Full clonotype sequence including the V gene, the amino acid CDR3 sequence and the J gene
 
 - VJ: V-J gene combination using V and J gene names

 - score: the alignment score communicated by the alignment tool

 - count: the occurrence of the clonotype, i.e the clone sequence

Clonotype tables must only contain a single chain. No paired-chain analysis are provided by the AnalyzAIRR package.
Pre-filtered files obtained using the \code{\link{filterClonotypes}} function can be used as input.
}
\examples{
l <- list.files(system.file(file.path('extdata/mixcr'),
                     package = 'AnalyzAIRR'),
                     full.names = TRUE)

dataset <- readAIRR(path = l[1],
                          fileFormat = "MiXCR",
                          chain = "TRA",
                          keep.ambiguous = FALSE,
                          keep.unproductive = FALSE,
                          outFiltered = FALSE,
                          aa.th = 8)

repseqexp_object <- RepSeqExp(clonotypetab = dataset)

}
\keyword{internal}
